Реализовано задание 3. Класс Exception404

+++ App/Controllers/Admin/Delete.php
@@ -3,6 +3,7 @@
 namespace App\Controllers\Admin;
 
 use App\ControllerAdmin;
+use App\Errors\Exception404;
 
 class Delete extends ControllerAdmin
 {
@@ -12,6 +13,10 @@ class Delete extends ControllerAdmin
         $params = $_GET;
         if (!empty($params['id'])){
             $article = \App\Models\Article::findById($params['id']);
+            if (!$article){
+                throw new Exception404('Нет такой новости.');
+            }
+
             $article->delete();
             header('Location: /admin/');
         }
